--keep, --overwrite

+++ trunk/md5merge.py
@@ -15,65 +15,85 @@ def largefileMD5( filename ):
 	return md5.hexdigest()
 
 def merge(dir1, dir2, relative_folder="."):	# recurse
+	global keep
+	global overwrite
+
 	print relative_folder+"/"
 
 	for item in os.listdir( dir2+"/"+relative_folder ):				# für jede Datei/jedes Verzeichnis, das "rechts" existiert ...
 
-		left_item = dir1+"/"+relative_folder+"/"+item
-		right_item = dir2+"/"+relative_folder+"/"+item
+		left_file = dir1+"/"+relative_folder+"/"+item
+		right_file = dir2+"/"+relative_folder+"/"+item
 
-		if os.path.isfile( right_item ):				#   ist es eine Datei:
-			if os.path.exists( left_item ):						# gibt es dieselbe Datei auch links?
-				right_filesize = os.path.getsize(right_item)				# Dateigröße rechts
-				left_filesize = os.path.getsize(left_item)				# Dateigröße links
-				if right_filesize != left_filesize:					# wenn die Dateigrößen schon ungleich sind, kann ich mir die MD5-Summe sparen
-					left_md5 = "Groessenunterschied!"
+		if os.path.isfile( right_file ):				#   ist es eine Datei:
+			if os.path.exists( left_file ):
+				left_filesize = os.path.getsize(left_file)
+				right_filesize = os.path.getsize(right_file)
+				if right_filesize != left_filesize:
+					left_md5 = "Size Mismatch"
+					right_md5 = "-"
 				else:
-					left_md5 = largefileMD5( left_item )
-				if left_filesize > 0:
-					right_md5 = largefileMD5( right_item )				# Prüfsumme rechts, die brauche ich auf jeden Fall
-				else:
-					right_md5 = "leere Datei ersetzen"
+					left_md5 = largefileMD5( left_file )
+					if left_filesize == 0:
+						right_md5 = "-"	# overwrite empty files left
+					else:
+						right_md5 = largefileMD5( right_file )
 			else:
-				left_md5 = "nicht vorhanden"						# ansonsten mach einen Dummy, der immer "ungleich" triggert
-				right_md5 = "nach links verschieben"
-			if left_md5 == right_md5:							# wenn die Prüfsummen gleich sind
-				os.remove( right_item )							# lösche die Datei rechts, belasse sie links
-				print ".",
-			else:										# wenn die Prüfsummen ungleich sind oder die Datei links nicht existiert
-				print "\n"+item+"\t\t\""+dir1+"\":"+left_md5+" , \""+dir2+"\":"+right_md5
-				if left_md5 == "nicht vorhanden":					# ... und benenne sie dabei um (um die linke nicht zu überschreiben)
-					newname = left_item
-				else:
-					newname = left_item+"-"+right_md5
-					if left_filesize == 0:						# wenn sie links leer ist, dann überschreibe sie einfach
-						os.remove( left_item )
-						newname = left_item
-				shutil.move( right_item, newname )
+				left_md5 = "Not Found"
+				right_md5 = "-"
 
-		elif os.path.isdir( right_item ):				# wenn es allerdings ein Verzeichnis ist
-			if not os.path.exists( left_item ):	
-				print "-> creating folder "+left_item
-				os.mkdir( left_item )						# erzeuge es ggf. auch links
-			merge( dir1, dir2, relative_folder=relative_folder+"/"+item )		# und führe die beiden Verzeichnisse zusammen
-			os.rmdir( right_item )							# nachdem das rechte danach leer sein sollte, können wir es löschen
+			if left_md5 == right_md5:	# files are equal
+				if not keep:
+					os.remove( right_file )
+				print ".",
+			else:				# not equal
+				print '\n'+item+'\t\t in "'+dir1+'": '+left_md5+' , in "'+dir2+'": '+right_md5
+				if left_md5 == "Not Found":
+					newname = left_file
+				else:
+					if overwrite:
+						os.remove( left_file )
+						newname = left_file
+					else:
+						if left_filesize == 0:
+							os.remove( left_file )
+							newname = left_file
+						else:
+							if right_md5 == "":
+								right_md5 = largefileMD5( right_file )
+							newname = left_file+"-"+right_md5
+				if keep:
+					shutil.copy( right_file, newname )
+				else:
+					shutil.move( right_file, newname )
+
+		elif os.path.isdir( right_file ):				# wenn es allerdings ein Verzeichnis ist
+			if not os.path.exists( left_file ):	
+				print "Creating Folder "+left_file
+				os.mkdir( left_file )						# erzeuge es ggf. auch links
+			merge( dir1, dir2, relative_folder=relative_folder+"/"+item )
+			if not keep:
+				os.rmdir( right_file )
 			print ""
 
 if len(sys.argv) < 3:
-	print " Dieses Skript führt zwei Verzeichnisse zusammen, die den gleichartigen Inhalt möglicherweise in unterschiedlichen Fassungen enthalten."
-	print " Alle Dateien werden anhand ihrer MD5-Prüfsummen verglichen."
-	print " Dateien, die sich unterscheiden, werden umbenannt und in das erstere angegebene Verzeichnis verschoben."
+	print "Usage: md5merge.py [--keep] [--overwrite] dir1 dir2 dir3 ..."
 	exit()
 
-opts, args = getopt.getopt(sys.argv[1:], "c", ["copy"])
-copy = False
+opts, args = getopt.getopt(sys.argv[1:], "k:o", ["keep", "overwrite"])
+global keep
+global overwrite
+keep = False
+overwrite = False
 for o, a in opts:
-	if o in ("-c", "--copy"):
-		copy = True
+	if o in ("-k", "--keep"):
+		keep = True
+	elif o in ("-o", "--overwrite"):
+		overwrite = True
 	else:
 		print 'Warning: Ignoring unrecognized argument "'+o+'".'
 
-dir1 = a[0]
+dir1 = args[0]
 if not os.path.exists( dir1 ):
 	print "Error: '"+dir1+"' not found"
 	sys.exit(1)
@@ -81,7 +101,7 @@ if not os.path.isdir( dir1 ):
 	print "Error: '"+dir1+"' is not a folder"
 	sys.exit(1)
 
-for dir2 in a[1:]:
+for dir2 in args[1:]:
 	if not os.path.exists( dir2 ):
 		print "Error: '"+dir2+"' not found"
 		break
@@ -89,8 +109,10 @@ for dir2 in a[1:]:
 		print "Error: '"+dir2+"' is not a folder"
 		break
 
-	print 'Merging "'+dir2+'" into "'+dir1+'" ..."
+	print 'Merging "'+dir2+'" into "'+dir1+'" ...'
 	merge(dir1, dir2)
-	print 'Removing remaining empty directory "'+dir2+'" ...'
-	os.rmdir(dir2)
+
+	if not keep:
+		print 'Removing remaining empty directory "'+dir2+'" ...'
+		os.rmdir(dir2)
